+++ src/extensions/ContentDiff/helpers/getters.js
@@ -1,0 +1,15 @@
+export const getId = (field) => {
+  return field.id;
+};
+
+export const getType = (field) => {
+  return field.type;
+};
+
+export const getLabel = (field) => {
+  return field.label;
+};
+
+export const getArrayType = (field) => {
+  return field.arrayType;
+};
